Eliminate a number of MSVC /W4 warnings (often just casting an unused parameter to void)
git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@1092 dfa36f5b-b2fa-0310-9605-e5aca7e64462

+++ Source/wali/nwa/details/SymbolStorage.hpp
@@ -525,7 +525,7 @@ namespace wali
     inline
     bool Label::containsSymbol( Sym sym ) const
     {
-      int count = syms.count(sym);
+      size_t count = syms.count(sym);
       if (neg) {
         //syms records symbols not on the edge, so
         //if the symbol is not in syms, it is in the collection
